Refactor ComparisonList into its own file.
